# specs/002-make-a-call/contracts/api-contracts.yaml
openapi: 3.1.0
info:
  title: promptalicious API
  description: API for LLM prompt execution and diagnostics
  version: 0.2.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
paths:
  /execute:
    post:
      operationId: executePrompt
      summary: Execute a system prompt against configured LLM
      description: |
        Sends the provided system prompt to the configured LLM (GPT-4o-mini initially)
        and returns the response with comprehensive diagnostic information.

        Functional Requirements: FR-002, FR-003, FR-005, FR-006, FR-009-014, FR-015-019
      tags:
        - Prompt Execution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutePromptRequest'
            examples:
              basicPrompt:
                summary: Basic system prompt
                value:
                  promptText: "You are a helpful assistant. Explain quantum computing in simple terms."
              longPrompt:
                summary: Longer technical prompt
                value:
                  promptText: "You are an expert code reviewer. Review the following TypeScript code for bugs, performance issues, and best practices: [code here]"
      responses:
        '200':
          description: Execution successful - returns result and diagnostics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutePromptSuccessResponse'
              examples:
                successfulExecution:
                  summary: Successful LLM response
                  value:
                    execution:
                      id: "550e8400-e29b-41d4-a716-446655440000"
                      promptText: "You are a helpful assistant. Explain quantum computing in simple terms."
                      executionTimestamp: "2025-11-04T10:30:00.000Z"
                      status: "completed"
                      targetModel: "gpt-4o-mini"
                    result:
                      id: "660e8400-e29b-41d4-a716-446655440001"
                      promptExecutionId: "550e8400-e29b-41d4-a716-446655440000"
                      responseText: "Quantum computing is a type of computing that uses quantum mechanics..."
                      inputTokenCount: 15
                      outputTokenCount: 120
                      totalTokenCount: 135
                      executionDurationMs: 1842
                      estimatedCostGBP: 0.0012
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutePromptErrorResponse'
              examples:
                emptyPrompt:
                  summary: Empty prompt text
                  value:
                    execution:
                      id: "770e8400-e29b-41d4-a716-446655440002"
                      promptText: ""
                      executionTimestamp: "2025-11-04T10:31:00.000Z"
                      status: "failed"
                      targetModel: "gpt-4o-mini"
                    error:
                      id: "880e8400-e29b-41d4-a716-446655440003"
                      promptExecutionId: "770e8400-e29b-41d4-a716-446655440002"
                      errorType: "validation"
                      errorMessage: "Prompt text cannot be empty"
                      timestamp: "2025-11-04T10:31:00.100Z"
        '401':
          description: Authentication error - invalid or missing API credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutePromptErrorResponse'
              examples:
                invalidApiKey:
                  summary: Invalid API key configured
                  value:
                    execution:
                      id: "990e8400-e29b-41d4-a716-446655440004"
                      promptText: "Test prompt"
                      executionTimestamp: "2025-11-04T10:32:00.000Z"
                      status: "failed"
                      targetModel: "gpt-4o-mini"
                    error:
                      id: "aa0e8400-e29b-41d4-a716-446655440005"
                      promptExecutionId: "990e8400-e29b-41d4-a716-446655440004"
                      errorType: "authentication"
                      errorCode: "invalid_api_key"
                      errorMessage: "The API key provided is invalid. Please check your configuration."
                      timestamp: "2025-11-04T10:32:00.200Z"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutePromptErrorResponse'
              examples:
                rateLimitExceeded:
                  summary: API rate limit hit
                  value:
                    execution:
                      id: "bb0e8400-e29b-41d4-a716-446655440006"
                      promptText: "Another prompt"
                      executionTimestamp: "2025-11-04T10:33:00.000Z"
                      status: "failed"
                      targetModel: "gpt-4o-mini"
                    error:
                      id: "cc0e8400-e29b-41d4-a716-446655440007"
                      promptExecutionId: "bb0e8400-e29b-41d4-a716-446655440006"
                      errorType: "rate_limit"
                      errorCode: "rate_limit_exceeded"
                      errorMessage: "Rate limit exceeded. Please try again in 60 seconds."
                      additionalContext:
                        retryAfter: 60
                      timestamp: "2025-11-04T10:33:00.300Z"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutePromptErrorResponse'
              examples:
                networkError:
                  summary: Network connectivity issue
                  value:
                    execution:
                      id: "dd0e8400-e29b-41d4-a716-446655440008"
                      promptText: "Yet another prompt"
                      executionTimestamp: "2025-11-04T10:34:00.000Z"
                      status: "failed"
                      targetModel: "gpt-4o-mini"
                    error:
                      id: "ee0e8400-e29b-41d4-a716-446655440009"
                      promptExecutionId: "dd0e8400-e29b-41d4-a716-446655440008"
                      errorType: "network"
                      errorMessage: "Unable to connect to OpenAI API. Please check your internet connection."
                      timestamp: "2025-11-04T10:34:00.400Z"
        '504':
          description: Gateway timeout - LLM call exceeded time limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutePromptErrorResponse'
              examples:
                timeoutError:
                  summary: Request timed out
                  value:
                    execution:
                      id: "ff0e8400-e29b-41d4-a716-446655440010"
                      promptText: "Timeout test prompt"
                      executionTimestamp: "2025-11-04T10:35:00.000Z"
                      status: "failed"
                      targetModel: "gpt-4o-mini"
                    error:
                      id: "000e8400-e29b-41d4-a716-446655440011"
                      promptExecutionId: "ff0e8400-e29b-41d4-a716-446655440010"
                      errorType: "timeout"
                      errorMessage: "The LLM request timed out after 60 seconds."
                      timestamp: "2025-11-04T10:36:00.000Z"
        '409':
          description: Conflict - execution already in progress
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                properties:
                  error:
                    type: object
                    required:
                      - errorType
                      - errorMessage
                    properties:
                      errorType:
                        type: string
                        enum: [conflict]
                      errorMessage:
                        type: string
                        example: "An execution is already in progress. Please wait for it to complete or cancel it before starting a new one."
              examples:
                executionInProgress:
                  summary: Concurrent execution attempted
                  value:
                    error:
                      errorType: "conflict"
                      errorMessage: "An execution is already in progress. Please wait for it to complete or cancel it before starting a new one."
...
  /execute/abort:
    post:
      operationId: abortExecution
      summary: Abort currently executing prompt
      description: |
        Aborts the currently executing prompt by triggering the AbortController signal.
        Returns success if execution was aborted, error if no execution in progress.

        Functional Requirements: FR-005a
      tags:
        - Prompt Execution
      responses:
        '200':
          description: Execution aborted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbortExecutionResponse'
              examples:
                abortSuccess:
                  summary: Successfully aborted
                  value:
                    success: true
                    message: "Execution aborted successfully"
        '400':
          description: No execution in progress to abort
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbortExecutionResponse'
              examples:
                noExecutionToAbort:
                  summary: Nothing to abort
                  value:
                    success: false
                    message: "No execution in progress to abort"
                    error:
                      errorType: "validation"
                      errorMessage: "No execution is currently in progress"
...
components:
  schemas:
    ExecutePromptRequest:
      type: object
      required:
        - promptText
      properties:
        promptText:
          type: string
          minLength: 1
          maxLength: 50000
          description: The system prompt to send to the LLM
          example: "You are a helpful assistant. Explain quantum computing in simple terms."
    ExecutePromptSuccessResponse:
      type: object
      required:
        - execution
        - result
      properties:
        execution:
          $ref: '#/components/schemas/PromptExecution'
        result:
          $ref: '#/components/schemas/ExecutionResult'
    ExecutePromptErrorResponse:
      type: object
      required:
        - execution
        - error
      properties:
        execution:
          $ref: '#/components/schemas/PromptExecution'
        error:
          $ref: '#/components/schemas/ExecutionError'
    PromptExecution:
      type: object
      required:
        - id
        - promptText
        - executionTimestamp
        - status
        - targetModel
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for this execution
        promptText:
          type: string
          description: The prompt that was executed
        executionTimestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when execution started
        status:
          type: string
          enum: [pending, in_progress, completed, failed, aborted]
          description: Current status of the execution
        targetModel:
          type: string
          description: The LLM model used
          example: "gpt-4o-mini"
    ExecutionResult:
      type: object
      required:
        - id
        - promptExecutionId
        - responseText
        - inputTokenCount
        - outputTokenCount
        - totalTokenCount
        - executionDurationMs
        - estimatedCostGBP
      properties:
        id:
          type: string
          format: uuid
        promptExecutionId:
          type: string
          format: uuid
        responseText:
          type: string
          description: The complete LLM response
        inputTokenCount:
          type: integer
          minimum: 1
          description: Number of tokens in the prompt
        outputTokenCount:
          type: integer
          minimum: 0
          description: Number of tokens in the response
        totalTokenCount:
          type: integer
          minimum: 1
          description: Sum of input and output tokens
        executionDurationMs:
          type: integer
          minimum: 1
          description: Time taken for the LLM call in milliseconds
        estimatedCostGBP:
          type: number
          format: decimal
          minimum: 0
          description: Estimated cost in British pounds (2 decimal places)
          example: 0.0012
    ExecutionError:
      type: object
      required:
        - id
        - promptExecutionId
        - errorType
        - errorMessage
        - timestamp
      properties:
        id:
          type: string
          format: uuid
        promptExecutionId:
          type: string
          format: uuid
        errorType:
          type: string
          enum: [authentication, network, api_error, timeout, rate_limit, validation, aborted, unknown]
          description: Classification of the error for debugging
        errorCode:
          type: string
          description: Provider-specific error code (optional)
          example: "insufficient_quota"
        errorMessage:
          type: string
          description: Human-readable error description
        stackTrace:
          type: string
          description: Technical stack trace (backend errors only, optional)
        additionalContext:
          type: object
          additionalProperties: true
          description: Structured error context (optional)
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when error occurred
...
    AbortExecutionResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          description: True if abort was successful
        message:
          type: string
          description: Human-readable result message
        error:
          type: object
          nullable: true
          properties:
            errorType:
              type: string
            errorMessage:
              type: string
...